Add MerkleAwareProvider visitors
This is a trick to avoid circular dependencies in the build. Whenever
we need to create a value declared in a module M that depends on core
(where MerkleAwareProvider is defined) we will have a circular
dependency since core now will have to import M. We cut the knot
by declaring a visitor producing the value, in the module where
the value type is declared.